[K/N] Don't use Xcode LLVM for bitcode compilation by default
Previously, when compiling for Apple targets, Kotlin/Native compiler
used Xcode LLVM to compile bitcode to machine code. There were multiple
reasons for that (mostly linked to bitcode embedding).

Now that those reasons are gone, it makes sense to use the LLVM version
bundled with the compiler for that purpose, as it always was with other
targets. This would give more flexibility in updating the bundled LLVM
version, because the compatibility between corresponding bitcode format
versions would no longer be necessary.

This commit performs this change by switching the default value for
`compileBitcodeWithXcodeLlvm` binary option to false.

^KT-59029