[LLVM 16] Fix anonymous/unnamed struct and enum support in cinterop
To distinguish structs and enums that have no name (and therefore need
one synthesized), cinterop uses a hack -- it checks that libclang names
for them contain "(anonymous" as a substring.

Between LLVM 11 and LLVM 16 this behavior is changed:
https://github.com/llvm/llvm-project/commit/50542d504dd869fe1241a9cc987b72ead5a56073
They are now "(unnamed".

Therefore, the current name-based check works wrong with LLVM 16.
This commit fixes this by using `clang_Cursor_isAnonymous` instead of
relying on names.

^KT-49279