[FIR] Unwrap typealiases inside FirTypeResolver
In the new test by @dimonchik0036, `@Deprecated` matches a similar
compiler-required annotation, so the compiler will try to resolve it at
`COMPILER_REQUIRED_ANNOTATIONS`. Even though it won't be saved as the
fqName doesn't match, the type resolver is instantiated with
`expandTypeAliases = true`, and it will fail due to requesting a higher
phase.

`COMPILER_REQUIRED_ANNOTATIONS` can't expand typealiases, so it should
instantiate the type resolver with `expandTypeAliases = false`. This was
previously missed because of the default value of `true`. This change
fixes `compiler/testData/diagnostics/tests/compilerRequiredAnnotationAsTypealias.kt`.