[K/WASM] Add a static initializers invoking phase to the IR lowerings ^KT-66103 Fixed
The `loweringList` in the WASM compiler currently lacks a phase dedicated to invoking static initializers.
During the `staticMembersLoweringPhase`, the companion object is extracted from the class without the corresponding initializer being invoked in the primary constructor.
Reusing the `invokeStaticInitializersPhase` from the JavaScript compiler and positioning it prior to the `staticMembersLoweringPhase` effectively resolves the issue.