[Gradle] Test for KLIB manifest in CompositeMetadataArtifactImpl
This should protect from cases when CompositeMetadataArtifact got
polluted due to some bugs or user misconfiguration.

One of the cases is when resource file is added to common source set.
In this case due to KT-66563 bug, resources will be mixed in to the
KLIB content. But if for some source set there is no sources.
Then in composite metadata artifact will be only the resources.
And previous implementation of CompositeMetadataArtifactImpl treated
that content as Metadata KLIB, however the metadata compiler didn't.
And thus it was failing with. This change makes it stricter.

Now it checks for `default/manifest` entry. This entry is considered
a sane check for KLIB content.

^KT-65315 Verification Pending