K2: don't report accidental override in case of final hidden method
Before this commit, we reported ACCIDENTAL_OVERRIDE_CLASH_BY_JVM_SIGNATURE
on an explicitly overridden function in case we could find
in the owner class scope a hidden function with the same signature.
For example: an explicit override fun get(Int) is declared,
a hidden charAt(Int) function, which is mapped to the same get(Int),
is implicitly overridden.

However, this situation creates in fact no problem in bytecode in case
a hidden conflicting method is final. See the test changed in this commit.

#KT-66463 In Progress