K2: Apply function type kind for FIR parameter expected type
For a value parameter, FIR resolution compares its actual type with the
expected type for a given argument expression. When the argument
expression has a function type, FIR resolution does not set the function
type kind for the expected type. As a result, we have a diagnostic error
reporting the mismatch between the actual and expected types. This
commit applies the function type kind for the expected type of the value
parameter.

^KT-66654 Fixed