K2 Scripting: fix return type calculation for script declarations
The ReturnTypeCalculatorWithJump.resolveDeclaration does not account
for the script as a declaration container, and it leads to the
resolution of the declaration in a wrong context (`withScript` is not
called). The implicitResolveWithJump.kts test demonstrates the problem.
It also leads to minor issue with inconsistent levels cfg levels
(see the changed destructuringDeclarationsScript.dot).
The fix adds script to the list of containers then needed.

#KT-66686 fixed
#KT-66352 fixed