JVM IR: refactor construction of existing lowering phases (3)
The goal here is to make each lowering phase a class, whose constructor
either takes one parameter of type BackendContext (JvmBackendContext in
case of JVM IR), or does not take any parameters. It simplifies the API
and allows the subsequent change where lowering phases can be referred
as just a constructor reference, with all meta-information (name,
description, prerequisites) available in an annotation on the class.