[FIR] Expand typealiases in supertypes
These are the tests that were broken initially by the absense of the
typealias in supertypes expansion.

In `assign` it's important to replace the attribute if exists: consider
types obtained via `typealias -> typealias -> class`. They must contain
the first typealias as the abbreviated type. For an example, see:
`FirPsiOldFrontendDiagnosticsTestGenerated.testKt15734`, where
`REDUNDANT_NULLABLE` disappears if the attribute is not replaced.

Checking `expandedTypeRef` is not enough to get the typealias expansion,
because the expanded type may have not yet been applied. See:
`FirPsiDiagnosticTestGenerated.testStatusResolveForTypealiasAsSuperClass`,
where `NOTHING_TO_OVERRIDE` appears otherwise.

Checking `FirImplicitBuiltinTypeRef` is needed to preserve things like
`FirImplicitAnyTypeRef`. For an example, see:
`FirOldFrontendMPPDiagnosticsWithPsiTestGenerated.testExpectAnnotationAndInlineClassWithDefaultValue`.