Fir2Ir: do not consider hidden declarations to be fake overrides
For example, see the test `codegen/box/specialBuiltins/charBuffer.kt`.
In that test, `CharBuffer.get` is hidden but it should not be considered
a fake override because it is actually declared in a Java class.

In a rare case, which will happen after enabling a workaround for
KT-65504, JVM backend would start to generate an incorrect bridge for
`KCB`, because bridge lowering will not be able to find a real
declaration for this "fake override", compute its JVM signature
incorrectly (with `get` instead of `charAt`) and add that incorrect
signature to the black list of final methods. Which will cause it to
generate an extra bridge `charAt`, which is bad because there's a final
`charAt` in the superclass.