JVM IR: refactor construction of existing lowering phases (1)
The goal here is to create subclasses in 'backend.jvm.lower' of those
lowering phases from 'backend.common' which either have a non-trivial
constructor, or have prerequisite phases declared in
'backend.jvm.lower'.

The reason is that in a subsequent commit, phase _variables_ (e.g.
`returnableBlocksPhase`) will be removed, phases will be referenced by
the constructor reference, and name/description/prerequisites will be
provided in an annotation on the class.