[Pill] Disable configuration cache for 'pill' and 'unpill'
As stated in the Gradle documentation, tasks that support the
configuration cache must not access information from 'Project's.
Instead, all required data needs to be passed explicitly via task
inputs.

As Pill analyzes the whole project structure, it is not realistic to
serialize it to some form and pass to the task.