[FIR] Commit inobviously false ABI changes
These are examples of tests where K2 loses information. Actually, these
are funny: K1 looses abbreviation altogether, and K2 renders the same
type instead of the abbreviation. This will be addressed in later
commits.

Also there are 2 more tests with `// JVM_ABI_K1_K2_DIFF: KT-63877`, but,
to be honest, they seem unrelated, because they just contain types that
are different in K1 and K2, not types that are equal to their own
typealias expansions or anything.

^KT-63877