Inference: Add some assertions instead of recently removed one
With PCLA, we've removed checkNotFixedTypeVariablesCountConsistency
at 276f5b26d8fe742eef2d76d8da6ab8b393678e64 since it was quite
hard to get rid of tones of false negatives because it's complicated
to compute `notFixedTypeVariablesToUse` properly.

But mostly the idea behind that assertion was to make sure that:
- we don't fix anything that should not be fixed yet (outer variables)
- there's nothing left unfixed

Thus, we introduced some quite obvious assertions here.

But we still leave KT-58742 open for some time in the hope that
in the future, we will figure out how to compute
`notFixedTypeVariablesToUse`.

^KT-66759 Related
^KT-58742 Open