[FIR] Improve tests for `REPEATED_ANNOTATION`/`@ParameterName`
In `duplicateParameterNameMinimized` calling `10.test { a ->`
results substitutes `Int` for `@ParameterName(foo) T`, resulting
in `@ParameterName(foo) Int`, but since it's the receiver of
`a.rest { b ->`, we do a second substitution of it into
`@ParameterName(bar) K`, thus obtaining
`b: @ParameterName(bar) @ParameterName(foo) Int`.
Previously, this example wouldn't work, because we'd have no
source on the return type ref of `b`, so we'd ignore the problem.

Also, after fixing the null sources,
`org.jetbrains.kotlin.konan.test.blackbox.FirStdlibTest` fails
with `initializationError`, because when compiling
`libraries/stdlib/test/collections/CollectionTest.kt`,
`CollectionTest::runningReduce` and specifically the call to
`runningReduce { acc, e -> acc + e })` inside it, we see that
the value parameter `acc` of this lambda has 2 instances of
`@R|kotlin/ParameterName|(name = String(index))` on its
return type `kotlin/Int`. Because this annotation is not
repeatable, we get an error and the test class fails.

And also for the very same reason
`:kotlin-stdlib:compileTestKotlinJvm` and
`:kotlin-stdlib:compileTestKotlinWasmJs` fail, but there it's
`libraries/stdlib/test/collections/CollectionTest.kt:443`