[FIR] Add a test to reproduce a regression introduced two commits before
The return type of a lambda whose last expression is a postfix
increment/decrement operator call on an array access would be
incorrectly inferred to Unit instead of the return type of the operator.

This is fixed in the next commit. More info about the reasons of this
regression can be found in that commit's message.