[FIR] Infer correct lambda return type when its last expr has no source
The `addReturnToLastStatementIfNeeded` function would set the generated
`return` expression's source to null if the source of the original
expression was null. This was the case for some synthetically generated
expressions, e.g. in desugared postfix increment/decrement
operators on arrays.

For example, the expression `a[i]++` is generated as follows:
```kotlin
{ // FirBlock
    val <array> = a
    val <index_0> = i
    val <unary> = <array>.get(<index_0>)
    <array>.set(<index_0>, <unary>.inc())
    <unary>
}
```

Here, the last expression `<unary>` didn't have a source,
so in a lambda a return expression with no source was generated.

Next, in `ControlFlowGraphBuilder#returnExpressionsOfAnonymousFunction`,
there is the following check:

```kotlin
lastStatement is FirReturnExpression &&
    lastStatement.target.labeledElement.symbol == function.symbol &&
    lastStatement.source?.kind != KtFakeSourceElementKind.ImplicitReturn.FromLastStatement
```

Basically, if this check is `true`, then `lastStatement` is ignored
when collecting the function's return expressions. This caused
this method to return an empty list of return expressions, which
in turn was the reason wjhy the lambda's return type was inferred to
Unit instead of whatever the return expression's type is.

Now, the interesting part is what happened when `lastStatement.source`
was `null` — the whole check was evaluated to `true`, which wasn't
right!

This is fixed by setting the return expression's `source` to that of
the corresponding function's body, if it otherwise is null.