[FIR] Prefer `var` properties as bases for intersection overrides
Picking a `var` avoids
`VAR_OVERRIDDEN_BY_VAL` reported for
intersection overrides. It's fine because
in cases where the code becomes green due
to not reporting this diagnostic, the user
is required to provide an explicit override
of such a property in a non-abstract subclass.

See: `compiler/testData/diagnostics/tests/varOverriddenByValThroughIntersection.kt`.

^KT-66717 Fixed