[FIR] Implement overridability check by checking either way in some cases
The cases are when JavaTypeParameterDefaultRepresentationWithDNN is
disabled and when the declarations are generic and either candidate
is from Java.
In these cases, the substitutor that is built from the type parameters,
either produces types with flexible upper bounds or not, depending
on the order of declarations.
This affects the result of AbstractTypeChecker.equalTypes while
JavaTypeParameterDefaultRepresentationWithDNN is disabled.

#KT-66701 Fixed