Tests: remove .kt.txt dumps in fake override tests
Tests in `fakeOverrides` are mainly supposed to test fake overrides and
their overridden symbols, which `.kt.txt` renderer does not display. So
unneeded dumps are removed to make it easier to work with these tests.

All tests got the `SKIP_KT_DUMP` directive, and two tests
(`intersectionOverride.kt`, `intersectionWithExplicitOverride.kt`) also
got `FIR_IDENTICAL` because the IR dumps (but not the `.kt.txt` dumps)
are the same.