[FIR] Ignore REPEATED_ANNOTATION on non-real sources
This commit basically reverts
"[FIR] Don't miss some existing REPEATED_ANNOTATIONS",
but was left here to avoid an unnecessary
and not-very-pleasing rebase.

This change makes sure stdlib is still
compilable with K2 after the previous changes.
Since no ultimate decision was accepted on
the matter, the compiler is allowed to
whatever it wants to, even ignore such
scenarios.

^KT-66990