[FIR] Check variable alias is stable before unwrapping
VariableStorage unwraps variable aliases using the provided Flow. This
can cause problems when the variable aliases can be reassigned in an
unstable way. Instead of automatically unwrapping all variable aliases,
first check that the variable is stable using the same rules as
smartcasting, and then unwrap the alias. Otherwise, don't apply any
statements to the variable.

^KT-66904 Fixed