[FIR] Consider when branches which check self type to be exhaustive
If a `when` expression contains a branch which checks for the exact or
super type of the subject, the `when` expression should be considered
exhaustive. The branch will have a `USELESS_IS_CHECK` warning, which may
help refactoring away from such patterns. This can help when migrating
from K1, as smartcasts in the subject of a `when` expression are only
applied in certain cases, while they are always applied in K2.

^KT-54920 Fixed