K2: Resolve FirProperty before access to initializer
When we resolve an initializer of a property, `toIrInitializer()`
can access the resolved type of a `FirLiteralExpression`. Since
`generateConstantExpressionByLiteral()` of `AbstractRawFirBuilder`
does not set type for some constants, the missing type causes an
exception for the type resolution. This commit runs FIR resolution to
set the type of `FirLiteralExpression` before accessing the resolved
type to avoid the exception.

^KT-67038 Fixed