[LLVM 16] Fix `platform.posix` build for mingw_x64 - add `htons`
The platform.posix lib definition for mingw_x64 uses `htons` function
which is not declared in the included headers, meaning that it relies
on the implicit declaration.

This becomes an error with Clang 16. To fix this, this commit adds an
explicit declaration for `htons` to `posix.def` for mingw_x64.

^KT-49279