[LLVM 16] Fix `target` attribute handling in cinterop for LLVM 16
Currently, the cinterop indexer looks for `__target__` attribute on the
functions to filter out those with non-trivial target.

This filtering is necessary because many of such functions are
`always_inline` (they can be found e.g. in `$llvm/lib/clang/*/include`),
so compiling a C wrapper for them fails with
```
always_inline function 'foo' requires target feature 'bar', but would be
inlined into function 'baz' that is compiled without support for 'bar'
```

Turns out that `target` attribute is equivalent to `__target__`: it does
the same thing, and parsed by Clang the same way (to `TargetAttr`).

This is not new in LLVM 16 (compared to LLVM 11). But headers provided
by Clang now use `target` extensively, which they didn't before.
See e.g. the following commit:
https://github.com/llvm/llvm-project/commit/b879f99f0ec0f34bfc5a6fe60299c308eec71f53

So, updating to LLVM 16 makes cinterop fail when compiling wrappers for
such functions with `target` attribute, because they are not filtered
out.

This commit fixes this by handling `target` attribute in cinterop
indexer the same way `__target__` is handled.

^KT-49279