[K/N][tests] Fix testObjCExport (noAutorelease) with -opt
1da8410c65f37cbef89557184fe2bfc821248c36 enabled optimizations for
Swift code in tests, when Kotlin code is compiled with optimizations.

This broke two noAutorelease tests. Apparently, when an object is
unused, Swift compiler can omit conversion from Objective-C to Swift
representation for that object. That's what happened in the test.

This commit fixes this by adding and using a 'blackHole' function,
to make sure that conversion code is always generated and tested.

^KT-65436