K2 IDE scripting: register assignment plugin if needed
Add custom compiler plugins registering for scripts in
order to correctly enable assignment plugin when appropriate
annotation is passed to the script's compiler options.
This is a temporary solution until we merge the new scripts
importing code to IJ, with which we'll be able to setup scripts'
modules correctly and thus trigger compiler plugins loading
by generic mechanisms (see #KT-67194).

#KT-65055 fixed