KT-65413 [LL FIR] Enable `switchableExtensionDeclarationsSymbolProvider` in `LLFirSourcesSession` for common modules
Without it, compiler plugins did not work well in common modules;
particularly, top-level `FirSymbolProvider` was not able to look up
classes generated by compiler plugins, and that cascaded to the frontend
being unable to resolve the callables from those generated classes.

N.B. The infrastructure for tests with common libraries dependencies for
AA is not ready yet, so we have to re-declare the annotations from
'plugin-annotations' module in the corresponding testdata to bypass
this limitation.

^KT-65413 Fixed