[LLVM 16] Fix undefined behavior for abs(Long/Int.MIN_VALUE)
`kotlin.math.abs` for K/N is implemented through standard C `*abs`
functions, e.g. `abs` and `llabs`.

The behavior of these functions is undefined for `INT_MIN` and
`LLONG_MIN`.
With LLVM 16, this started triggering failures for stdlib tests
`numbers.IntegerMathTest.intSigns` and `.longSigns` when compiling
with `-opt` on x86_64.

This commit fixes that by using `__builtin_elementwise_abs` instead of
`abs`/`llabs`. Both options are intrinsifed in LLVM, and differ only
in the behavior for `INT_MIN` and `LLONG_MIN`.

^KT-49279