[Gradle, Kapt] Remove the logic of adding objects to Groovy closure context
Particularly, the functions arguments(Closure) and javacOptions(Closure) should be autogenerated in runtime by Gradle to be accessible from Groovy from the related functions taking Action. The closures shall receive the proper context automatically.
The `getAdditionalArgumentsForJavac` method is unused for quite a while, so it scheduled for removal without replacements
^KT-61928 Fixed