Set correct IR origins for augmented assigment-originated calls
Apparently, there is no difference between a[b] += c and a += c in terms
of origin calculation, thus `DesugaredArrayPlusAssign` and similar ones
have been renamed to `DesugaredPlusAssign` and reused for non-array
cases. `DesugaredCompoundAssignment` becomes redundant as it covered
by these fake source kinds and the one special one, namely
`AssignmentPluginAltered` which is used in
`FirAssignmentPluginAssignAltererExtension`.

^KT-60261: Fixed