JVM: Extend range of `this` in suspend lambda
to cover the whole function. It is safe in case of
suspend lambdas, since their `this` is continuation,
which is always present. The change is required to
step through completion chain in the debugger, thus
it is required for stepping. We have already done
so for $completion parameter of suspend functions,
see f1fd84f539ca5f5b314d0f1ce83c89ad94d26319.

 #KT-67136 Fixed