[gradle, mpp] Add test that checks DR between host-specific source sets
On of the previous commits made it so that the klibArtifacts are created
even if the target is not supported on the current host.

The fact that they were not created was actually an issue, as without
artifacts the dependencies between such source sets couldn't be resolved.

This is a minor bug, because:
a) it concerns UX of working with Darwin-targets on non-Mac host, which
   was always "best effort"
b) you need to write code in leaf Darwin source-set, which is a very rare
   occassion (source sets like `iosArm64` almost never contain any source,
   instead, people write code in `ios`)

See https://jetbrains.team/p/kt/reviews/15211