[K2] Stop reporting NO_ACTUAL_CLASS_MEMBER_FOR_EXPECTED_CLASS when there are more than one matching actuals
^KT-66723 Fixed
Review: https://jetbrains.team/p/kt/reviews/15462

NO_ACTUAL_CLASS_MEMBER_FOR_EXPECTED_CLASS is a frontend error.
AMBIGUOUS_ACTUALS is a backend error. If we report a error on frontend
than we never get a chance to report error on backend

The commit improves the error reporting, because previously it wasn't
obvious that AMBIGUOUS_ACTUALS is the real problem.