[FIR] Resolve foreign annotations coming from typealias expansions
This change is backed by
`analysis/low-level-api-fir/testData/lazyResolveTypeAnnotations/`
`typeAlias/insideFunction.kt`. Without the change the test would fail:

> Caused by: org.jetbrains.kotlin.utils.exceptions
> .KotlinIllegalStateExceptionWithAttachments: Expected
> FirResolvedArgumentList for FirAnnotationCallImpl of
> FirValueParameterImpl(Source) but FirArgumentListImpl found

The test can be simplified as:

```
typealias MyTypeAlias = @TypeAnnWithArg("a") Int
fun resolveMe(val param: @TypeAnnWithArg("b") MyTypeAlias) {}
```

It fails, because after resolving with
`LLFirAnnotationArgumentsTargetResolver` we call `checkIsResolved()`
and it traverses typeref's custom annotations, thus visiting both
`@TypeAnnWithArg("a")` and `@TypeAnnWithArg("b")`, but only the latter
is properly resolved. This is because foreign annotations are ignored as
`return annotationCall` in `FirAnnotationArgumentsTransformer`.

This change makes it so that foreign annotations are collected from
`resolveMe()` as well and later resolved properly.

The change in `ForeignAnnotationsCollector` is needed to avoid postponing
erroneous annotations: these are created for some annotations on type
parameters. See:
`LLFirPreresolvedReversedDiagnosticCompilerFE10TestDataTestGenerated`
`.Tests#testKt46483`.

This change is also backed by:
`LLFirPreresolvedReversedDiagnosticCompilerFE10TestDataTestGenerated`
`.Tests.Typealias#testTypealiasRhsRepeatedAnnotationInArguments`.
Otherwise, produces:

> Caused by: org.jetbrains.kotlin.utils.exceptions
> .KotlinIllegalStateExceptionWithAttachments: Expected
> FirResolvedArgumentList for FirAnnotationCallImpl of
> FirTypeAliasImpl(Source) but FirArgumentListImpl found`

...because the typealias declaration itself can contain foreign
annotations in its expansion, so it must support postponing as well).

Also, fix owners of error annotation calls and add a test that ensures
PSI and LT produce the same owner. `FirAnnotationCall`s'
`containingDeclarationSymbol` don't refer to the corresponding type
parameter, but rather to their owners. For light tree
`FirErrorAnnotationCall`s seem to have gotten it right, but for PSI they
have not.