[JPS] Disable processing of removed classes in JPS graph implementation
How it worked before: We run the compiler to build psi from dirty files before compilation round and extract existing class declarations from it. Then we compare the current ones with cached declarations from the classToOutput map to find removed classes. We get their fqname and map on lookups to expand the scope with affected files before the main compiler launch.
Motivation: This approach does not work with the new JPS graph implementation because we don't have the own lookup storage anymore.
How it will work now: This logic will be replaced with expanding the compilation scope on the JPS graph side.

Also, this change will help to reduce time spent on loading caches and creating environment, especially on modules without Kotlin files

Relates to KT-65043