[IR] fail fast refactoring: expect and actual must always have the same IR node
Motivation: while I was working on KT-66436 one of the prototypes I had
actualized expect property with field in Java (those are different
IrSymbol nodes).

But my actualization was silently ignored because safe-cast `as? S` was
returning null, so I had to spend more time debugging before I could
discover that nodes must have the same IrSymbol type.

It's weird to ignore what's written in the mapping if it's a different
type. It's better to fail fast to manifest the problem in the mapping.