[Scripting] Copy result property when copying IrScript nodes
#KT-67066 Fixed

IrScript was not including the result property when the node was being copied. This would result in custom scripting implementations returning Unit as the return value from scripts rather than the expected result.

No test is included as it requires a custom compiler plugin and the test setup isn't setup to easily add this.


Merge-request: KT-MR-15526
Merged-by: Christian Melchior <christian.melchior@jetbrains.com>