[FIR] Fix metadata serialization of AbbreviatedType
`preserveOriginalType` is neeeded to prevent the typealias type from
being "broken down" further: as it happens with suspend function types
for example. In such cases the `typeOrTypealiasProto()` delegates to
`typeProto()`, but with a different type, thus the abbreviation info is
lost.

^KT-63877
^KT-68087

The change is also backed by:
`FirLightTreeBlackBoxCodegenTestGenerated.Multiplatform.K2`
`#testAliasSuperTypeInLazy`. Otherwise:

> Caused by: java.lang.IllegalStateException: IR class for BaseAlias not
> found

When the supertypes of the common `/Child` are expanded, the typealias
is fully expanded with the use site session of `/Child`, thus resulting
in `expect class BaseAlias`. When the `FirElementSerializer` serializes
types, it expands them with the use site session, thus we now need to
expand typealiases further.