[LLVM 16] Update LLVM link options
To interact with LLVM as a library, K/N compiler links many LLVM
components as static libraries into a single JNI dynamic library.
The exact set of required LLVM static libraries, as well as the order,
depends on the LLVM version.

This commit updates them (encoded in the form of linker flags) to match
LLVM 16.

^KT-49279