[LLVM 16] Opaque pointers: fix checking return object types
Currently, for the Kotlin/Native code generator, it matters whether a
function returns object reference or not. The compiler checks LLVM
function return types for that. But with LLVM opaque pointers, it won't
be possible to distinguish object reference from another pointer type.

This commit addresses that by using IR types or other contextual
information for such checks and storing that in `LlvmRetType` and
`LlvmCallable` whenever necessary.

^KT-67692

Co-authored-by: Svyatoslav Scherbina <svyatoslav.scherbina@jetbrains.com>