[tree generators] Get rid of context receivers
In the tree generators, we used context receivers to pass around
`ImportCollector`. We want to drop context receivers because
this is an experimental feature that will soon be replaced with
context parameters. Before that, all usages of context receivers
must be eliminated from the compiler code.

Here, since most of the times `ImportCollector` comes together with
`SmartPrinter`, we replace those two with a single
`ImportCollectingPrinter` class.

^KT-67129 Fixed