fixup! [FIR] Expand typealiases in supertypes
Recursive expansion is needed, because otherwise we can't call
`fullyExpandedType()` as it wont have all the needed data.

The change in `getResolvedExpandedType()` is needed, because
`typeAlias.expandedConeType` returns `null` if the typealias expands to
something incorrect, like its own type parameter. In this case, it's
already resolved, so we don't need to look up anything in the session.

This change is backed by:
`LLFirPreresolvedReversedDiagnosticCompilerFE10TestDataTestGenerated`
`.Tests.Typealias#testTypeAliasShouldExpandToClass`. Otherwise:

> java.lang.IllegalArgumentException: Unexpected status at
> FirApplySupertypesTransformer: org.jetbrains.kotlin.fir.resolve.transformers
> .SupertypeComputationStatus$NotComputed@3839fdcb for
> /Outer.ToTypeParam1

...because non-recursive expansion of a typealias pointing to
`Outer.ToTypeParam1` leads to our inability to fully expand
`Outer.toTypeParam1`.