[LLVM 16] Opaque pointers: fix generating TypeInfo
Currently, when the Kotlin/Native compiler generates a TypeInfo, it
has to know which fields are object references (the GC uses that info).

The compiler checks field LLVM types for that. But with LLVM opaque
pointers, it won't be possible to distinguish object reference from
another pointer type.

This commit addresses that by using IR types or other contextual
information for such checks and storing that in `ClassLlvmDeclarations`.

^KT-67692

Co-authored-by: Svyatoslav Scherbina <svyatoslav.scherbina@jetbrains.com>