[LLVM 16] Opaque pointers: fix generating memory accesses
Currently, when the Kotlin/Native compiler generates memory accesses,
it has to know which accesses handle object references (to properly
involve the GC for them).

The compiler checks LLVM types for that. But with LLVM opaque pointers,
it won't be possible to distinguish object reference from another
pointer type.

This commit addresses that by using IR types or other contextual
information for such checks.

^KT-67692

Co-authored-by: Svyatoslav Scherbina <svyatoslav.scherbina@jetbrains.com>