[LLVM 16] Opaque pointers: fix generating thread locals
Currently, when the Kotlin/Native compiler generates thread locals,
it has to know which declarations are object reference typed
(to properly support the GC for them).

The compiler checks LLVM types for that. But with LLVM opaque pointers,
it won't be possible to distinguish object reference from another
pointer type.

This commit addresses that by using IR types or other contextual
information for such checks.

^KT-67692