Support reporting of PACKAGE_REDECLARATION in K2
In K1 it's reported on both sides - file package and redeclaring
class. I decided to implement reporting only from the package side
because:

1. There is an implementation obstacle. I tried to use
`FirProvider#getFirFilesByPackage` to find files with the same package
as classId of current class, and it's not supported in `LLFirProvider`
and just throws an exception. It could be implemented, actually, unless
there were other arguments "against" below.
2. This looks like a duplication these two reports always go
together. So, removing one of doesn't turn any red code into green.
3. If we go the other way around, and we have only a reporting from
class side, then we can't detect cases when there is `package a.b` and
`class a`.
4. Less code, more strict type parameters of diagnostic factory
(KtPackageDirective+ClassId instead of KtElement+String).

^KT-59897