Enable compose compiler plugin for CodeGen test
We recently added the compose compiler plugin to KT compiler repo.
Before we added it, we mimicked the plugin to reproduce compiler bugs
triggered by the plugin. As time goes, we copied more plugin code to KT
compiler repo, which is not good for maintenance. Since we have the
plugin as a part of KT compiler now, we can enable it for CodeGen test.

Note that the compose compiler plugin has strong dependency on compose
runtime, this commit adds compose-runtime-desktop jar file (otherwise,
we have to write more than 20 mock classes for the compose runtime).