[Gradle] Try to filter the same klib for JS test compilation
Some projects may bring as transitive dependency same klib that is
produced by main compilation in the exploded manner (unzipped klib)
which leads to compiler warnings on same klibs found in inputs.

This fix is only a workaround for this specific problem with JS test
compilations. Proper solution should be worked on in the scope of
KT-65285.

^KT-64115 Verification Pending