[LLVM 16] Fix __ENVIRONMENT_OS_VERSION_MIN_REQUIRED__ runtime warning
When compiling K/N runtime after the update to LLVM 16, C++ compiler
emits compiler warnings about `__ENVIRONMENT_OS_VERSION_MIN_REQUIRED__`
macro being redefined.

The reason is that this macro is defined in `ClangArgs` (to workaround
the lack of it in Clang 11), but also defined by Clang 16. But removing
the workaround from `ClangArgs` wouldn't help right away, because the
runtime build scripts are using the bootstrap version of `ClangArgs`.

Therefore, we have to temporary filter the problematic flag in the
build scripts, until the bootstrap is updated after the LLVM update.

^KT-67730