fixup! [FIR] Fix metadata serialization of AbbreviatedType
Backed by:
`FirLightTreeBlackBoxCodegenTestGenerated.Multiplatform.K2`
`#testAliasSuperTypeInLazy`. Otherwise:

> Caused by: java.lang.IllegalStateException: IR class for BaseAlias not
> found

When the supertypes of the common `/Child` are expanded, the typealias
is fully expanded with the use site session of `/Child`, thus resulting
in `expect class BaseAlias`. When the `FirElementSerializer` serializes
types, it expands them with the use site session, thus we now need to
expand typealiases further.