[K2] Try to use `evaluatedInitializer` inside lazy FIR to IR declarations
In case of source dependency (as in
`FirPsiJsOldFrontendDiagnosticsWithBackendTestGenerated`
tests) the compiler is not going to create binary artifact with
constant values in constant properties. Instead, the compiler
is going to use FIR initializer as it was. That means that constant
`IrProperty` from dependant module could miss its constant initializer.
This breaks up IR interpreter logic, and the solution is to use
`evaluatedInitializer` to create IR initializer.

#KT-62304 Fixed