Enable compose compiler plugin for binary library used by compiler tests
If we enable the compose compiler plugin for the source module of
compiler (or CodeGen) tests, lower transformers of the plugin will be
skipped for binary libraries, which results in mismatch of signature
between function caller and function. For example, the compose compose
plugin adds `androidx.compose.runtime.Composer` parameter to the
function with `@Composable` annotation. Without this change, only the
source module will be transformed. Therefore, the addition parameter
causes function signature mismatch.