fixup! [FIR] Commit other test improvements
Initially these tests were broken by my changes and required a fix in
`AbbreviatedTypeAttribute`, but after having rebased on a fresher master
they are now fine. This is thanks to a similar fix by Kirill in
`c3926219`.

These tests rely on the way we compare types with and without
`AbbreviatedTypeAttribute`: we must consider them equal (`equals()`).

This commit commits the false changes, and the next one reverts them
(commits the "fix" or where it used to be). This way it's easier to
track than just describe these tests in the commit message of the
previous commit.