[K/N][build] Set macOS deployment version for LLVM to 11.2
This commit changes the deployment version for LLVM binaries to 11.2
on macOS.

By default, the deployment version is set to the OS version of the
build agent. So this change is important to make sure that the built
LLVM is runnable on older macOS versions.

11.2 is the version used for the LLVM 11 build currently used by Kotlin.

^KT-70566

(cherry picked from commit 24c70d705e4b0c2b84baeb3424b313c8ef2ac91c)