KT-67701 [LL FIR] Introduce separate caches for functions and properties in `LLFirProviderHelper`
To compute `getTopLevelFunctionSymbols` and
`getTopLevelPropertySymbols`, different sets of files
were passed by `LLFirCombinedKotlinSymbolProvider`

When shared `callablesByCallableId` was used, it could cache the value
for a single context, and then return it for the other contexts also.

Since callable declarations can have identical `CallableId`s
(a property and a function, for example), this behavior led to
incorrect resolve later on.

Now, with the separate caches for functions and properties, this
problem should be mitigated.

Note that the caches still do not remember the contexts used for the
computation, but it is the responsibility of the `LLFirProviderHelper`'s
user to always pass the "full" and complete context to all its members.
If this statement holds, then it should not matter whether the
cache value was computed with the context or without it - they should
be equal.

^KT-67701 Fixed