[FIR] Always render abbreviations in analysis tests
There's already a place where we do (`KtAnalysisSession.renderType`), so
we should be consistent. I believe it's way better always to render them
to avoid missing typealias expansion issues early.

Without this change the attached test produces conflicting changes that
can't be committed (trying to compare renderings of the same symbol with
and without the abbreviation).